+++ dataclass_wizard/class_helper.py
@@ -3,9 +3,11 @@ from dataclasses import Field, fields
 from typing import Dict, Tuple, Type, Union, Callable, Optional
 
 from .abstractions import W, AbstractLoader, AbstractDumper
+from .models import JSONField, JSON
 from .parsers import Parser
 from .type_def import ExplicitNullType, T
 from .utils.dict_helper import CaseInsensitiveDict
+from .utils.type_check import is_annotated, get_args
 
 
 # A cached mapping of fully qualified class name to the list of fields, as
@@ -23,6 +25,11 @@ _CLASS_TO_DUMPER: Dict[str, Type[AbstractDumper]] = {}
 #
 # Note: need to create a `ForwardRef` here, because Python 3.6 complains.
 _FIELD_NAME_TO_LOAD_PARSER: Dict[str, 'CaseInsensitiveDict[str, Parser]'] = {}
+
+# Since the dump process doesn't use Parsers currently, we use a sentinel
+# mapping to confirm if we need to setup the dump config for a dataclass
+# on an initial run.
+_IS_DUMP_CONFIG_SETUP: Dict[str, bool] = {}
 
 # A cached mapping, per dataclass, of JSON field to instance field name
 _JSON_FIELD_TO_DATACLASS_FIELD: Dict[
@@ -98,13 +105,108 @@ def dataclass_field_to_load_parser(
     name = get_class_name(cls)
 
     if name not in _FIELD_NAME_TO_LOAD_PARSER:
-        # Lookup the Parser (dispatcher) for each field based on its annotated
-        # type, and then cache it so we don't need to lookup each time.
-        _FIELD_NAME_TO_LOAD_PARSER[name] = CaseInsensitiveDict(
-            {f.name: cls_loader.get_parser_for_annotation(f.type, cls)
-             for f in dataclass_fields(cls)})
+        _setup_load_config_for_cls(cls_loader, cls, name)
 
     return _FIELD_NAME_TO_LOAD_PARSER[name]
+
+
+def _setup_load_config_for_cls(cls_loader, cls, name: str):
+    """
+    This function processes a class `cls` on an initial run, and sets up the
+    load process for `cls` by iterating over each dataclass field. For each
+    field, it performs the following tasks:
+
+        * Lookup the Parser (dispatcher) for the field based on its type
+          annotation, and then cache it so we don't need to lookup each time.
+
+        * Check if the field's annotation is of type ``Annotated``. If so,
+          we iterate over each ``Annotated`` argument and find any special
+          :class:`JSON` objects (this can also be set via the helper function
+          ``json_key``). Assuming we find it, the class-specific mapping of
+          JSON key to dataclass field name is then updated with the input
+          passed in to this object.
+
+        * Check if the field type is a :class:`JSONField` object (this can
+          also be set by the helper function ``json_field``). Assuming this is
+          the case, the class-specific mapping of JSON key to dataclass field
+          name is then updated with the input passed in to the :class:`JSON`
+          attribute.
+    """
+
+    json_to_dataclass_field = _JSON_FIELD_TO_DATACLASS_FIELD[name]
+    name_to_parser = {}
+
+    for f in dataclass_fields(cls):
+
+        # Lookup the Parser (dispatcher) for each field based on its annotated
+        # type, and then cache it so we don't need to lookup each time.
+        name_to_parser[f.name] = cls_loader.get_parser_for_annotation(
+            f.type, cls)
+
+        # Check if the field is a `JSONField` object. If so, update the
+        # class-specific mapping of JSON key to dataclass field name.
+        if isinstance(f, JSONField):
+            for key in f.json_key.keys:
+                json_to_dataclass_field[key] = f.name
+
+        # Check if the field annotation is an `Annotated` type. If so,
+        # look for any `JSON` objects in the arguments; for each object,
+        # update the class-specific mapping of JSON key to dataclass field
+        # name.
+        if is_annotated(f.type):
+            for extra in get_args(f.type)[1:]:
+                if isinstance(extra, JSON):
+                    for key in extra.keys:
+                        json_to_dataclass_field[key] = f.name
+
+    _FIELD_NAME_TO_LOAD_PARSER[name] = CaseInsensitiveDict(name_to_parser)
+
+
+def setup_dump_config_for_cls_if_needed(cls):
+    """
+    This function processes a class `cls` on an initial run, and sets up the
+    dump process for `cls` by iterating over each dataclass field. For each
+    field, it performs the following tasks:
+
+        * Check if the field's annotation is of type ``Annotated``. If so,
+          we iterate over each ``Annotated`` argument and find any special
+          :class:`JSON` objects (this can also be set via the helper function
+          ``json_key``). Assuming we find it, the class-specific mapping of
+          dataclass field name to JSON key is then updated with the input
+          passed in to this object.
+
+        * Check if the field type is a :class:`JSONField` object (this can
+          also be set by the helper function ``json_field``). Assuming this is
+          the case, the class-specific mapping of dataclass field name to JSON
+          key is then updated with the input passed in to the :class:`JSON`
+          attribute.
+    """
+
+    name = get_class_name(cls)
+
+    if name in _IS_DUMP_CONFIG_SETUP:
+        return
+
+    dataclass_to_json_field = _DATACLASS_FIELD_TO_JSON_FIELD[name]
+
+    for f in dataclass_fields(cls):
+
+        # Check if the field is a `JSONField` object. If so, update the
+        # class-specific mapping of dataclass field name to JSON key.
+        if isinstance(f, JSONField) and f.json_key.all:
+            dataclass_to_json_field[f.name] = f.json_key.keys[0]
+
+        # Check if the field annotation is an `Annotated` type. If so,
+        # look for any `JSON` objects in the arguments; for each object,
+        # update the class-specific mapping of dataclass field name to JSON
+        # key.
+        if is_annotated(f.type):
+            for extra in get_args(f.type)[1:]:
+                if isinstance(extra, JSON) and extra.all:
+                    dataclass_to_json_field[f.name] = extra.keys[0]
+
+    # Mark the dataclass as processed, as the initial dump process is set up.
+    _IS_DUMP_CONFIG_SETUP[name] = True
 
 
 def call_meta_initializer_if_needed(cls: Type[W]):