+++ tests/unit/test_load.py
@@ -7,11 +7,14 @@ import logging
 from collections import namedtuple, defaultdict
 from dataclasses import dataclass
 from datetime import datetime, date, time
-from typing import List, Optional, Union, Tuple, Dict, NamedTuple, Type, DefaultDict
+from typing import (
+    List, Optional, Union, Tuple, Dict, NamedTuple, Type, DefaultDict,
+    Set, FrozenSet
+)
 
 import pytest
 
-from dataclass_wizard import JSONSerializable
+from dataclass_wizard import JSONSerializable, json_field, json_key
 from dataclass_wizard.errors import ParseError
 from dataclass_wizard.parsers import OptionalParser, Parser
 from dataclass_wizard.type_def import NoneType, T
@@ -48,6 +51,104 @@ def test_bool(input, expected):
     log.debug('Parsed object: %r', result)
 
     assert result.my_bool == expected
+
+
+def test_from_dict_key_transform_with_json_field():
+    """
+    Specifying a custom mapping of JSON key to dataclass field, via the
+    `json_field` helper function.
+    """
+
+    @dataclass
+    class MyClass(JSONSerializable):
+        my_str: str = json_field('myCustomStr')
+        my_bool: bool = json_field(('my_json_bool', 'myTestBool'))
+
+    value = 'Testing'
+    d = {'myCustomStr': value, 'myTestBool': 'true'}
+
+    result = MyClass.from_dict(d)
+    log.debug('Parsed object: %r', result)
+
+    assert result.my_str == value
+    assert result.my_bool is True
+
+
+def test_from_dict_key_transform_with_json_key():
+    """
+    Specifying a custom mapping of JSON key to dataclass field, via the
+    `json_key` helper function.
+    """
+
+    @dataclass
+    class MyClass(JSONSerializable):
+        my_str: Annotated[str, json_key('myCustomStr')]
+        my_bool: Annotated[bool, json_key('my_json_bool', 'myTestBool')]
+
+    value = 'Testing'
+    d = {'myCustomStr': value, 'myTestBool': 'true'}
+
+    result = MyClass.from_dict(d)
+    log.debug('Parsed object: %r', result)
+
+    assert result.my_str == value
+    assert result.my_bool is True
+
+
+@pytest.mark.parametrize(
+    'input,expected,expectation',
+    [
+        ([1, '2', 3], {1, 2, 3}, does_not_raise()),
+        ('TrUe', True, pytest.raises(ParseError)),
+        ((3.22, 2.11, 1.22), {3, 2, 1}, does_not_raise()),
+    ]
+)
+def test_set(input, expected, expectation):
+
+    @dataclass
+    class MyClass(JSONSerializable):
+        num_set: Set[int]
+        any_set: set
+
+    d = {'numSet': input, 'any_set': input}
+
+    with expectation:
+        result = MyClass.from_dict(d)
+        log.debug('Parsed object: %r', result)
+
+        assert isinstance(result.num_set, set)
+        assert isinstance(result.any_set, set)
+
+        assert result.num_set == expected
+        assert result.any_set == set(input)
+
+
+@pytest.mark.parametrize(
+    'input,expected,expectation',
+    [
+        ([1, '2', 3], {1, 2, 3}, does_not_raise()),
+        ('TrUe', True, pytest.raises(ParseError)),
+        ((3.22, 2.11, 1.22), {1, 2, 3}, does_not_raise()),
+    ]
+)
+def test_frozenset(input, expected, expectation):
+
+    @dataclass
+    class MyClass(JSONSerializable):
+        num_set: FrozenSet[int]
+        any_set: frozenset
+
+    d = {'numSet': input, 'any_set': input}
+
+    with expectation:
+        result = MyClass.from_dict(d)
+        log.debug('Parsed object: %r', result)
+
+        assert isinstance(result.num_set, frozenset)
+        assert isinstance(result.any_set, frozenset)
+
+        assert result.num_set == expected
+        assert result.any_set == frozenset(input)
 
 
 @pytest.mark.parametrize(
@@ -920,3 +1021,16 @@ def test_optional_parser_contains(input, expected):
 
     actual = input in optional_parser
     assert actual == expected
+
+
+def test_parser_with_unsupported_type():
+    """
+    Test case for :meth:`Parser.__call__` with an unknown or unsupported
+    type, added for code coverage.
+
+    """
+    base_type: Type[T] = str
+    mock_parser = Parser(None, base_type, hook=None)
+
+    with pytest.raises(ParseError):
+        _ = mock_parser('hello world')